==> CLEAN CICFLOWMETER/01. Data Preparation/00. Merge Distributed Dataset.ipynb <==
import pandas as pd
from glob import glob

from imblearn.under_sampling import RandomUnderSampler

from sklearn import preprocessing

def get_csv(uri):
    csvs = []
    if "*" in uri:
        all_csv = glob(uri)
        [ csvs.append(pd.read_csv(uri)) for uri in all_csv ]
        return pd.concat(csvs)
    return pd.read_csv(uri)

dropped_features = [
    "Flow ID", "Src IP", "Src Port",
    "Dst IP", "Dst Port", "Timestamp"
]

def label_dataset(ds, label):
    dataset = ds.copy()
    dataset[["Label"]] = label
    return dataset

def remove_string(ds):
    # Hapus fitur yang memiliki string --------------------------------
    print("Jumlah fitur sebelum didrop: " + str(ds.shape[1]))
    ds.drop(columns=dropped_features, inplace=True)
    print("Jumlah fitur setelah didrop: " + str(ds.shape[1]))

### Malware IoT-23

malware = get_csv('/media/kmdr7/Seagate/DATASETS/IOT-23/CTU-IoT-Malware-Capture-48-1/out2/*')
malware = label_dataset(malware, "Malicious")
remove_string(malware)

### Benign IoT-23

benign_iot23 = get_csv('/media/kmdr7/Seagate/DATASETS/IoT-Traffic-Traces/out/*')
benign_iot23 = label_dataset(benign_iot23, "Benign")
remove_string(benign_iot23)

### Honeypot IoT-23

honeypot = pd.concat([
    get_csv('/media/kmdr7/Seagate/DATASETS/IOT-23/CTU-Honeypot-Capture-7-1/out/*.csv'),
    get_csv('/media/kmdr7/Seagate/DATASETS/IOT-23/CTU-Honeypot-Capture-4-1/out/*.csv'),
    get_csv('/media/kmdr7/Seagate/DATASETS/IOT-23/CTU-Honeypot-Capture-5-1/out/*.csv')
])
honeypot = label_dataset(honeypot, "Benign")
remove_string(honeypot)

### Benign IoT-Benign-Trafic

# benign_iottr = get_csv('/media/kmdr7/Seagate/DATASETS/IoT-Benign-Traffic/out/18-10-17.pcap_Flow.csv')
# honeypot = pd.concat([
#     get_csv('/media/kmdr7/Seagate/DATASETS/IoT-Benign-Traffic/out/CTU-Honeypot-Capture-7-1/out/*.csv'),
#     get_csv('/media/kmdr7/Seagate/DATASETS/IoT-Benign-Traffic/out/CTU-Honeypot-Capture-4-1/out/*.csv'),
#     get_csv('/media/kmdr7/Seagate/DATASETS/IoT-Benign-Traffic/out/CTU-Honeypot-Capture-5-1/out/*.csv')
# ])
# benign_iottr = label_dataset(benign_iottr, "Benign")
# remove_string(benign_iottr)

print("Malware Shape " + str(malware.shape))
print("Benign IoT-23 Shape " + str(benign_iot23.shape))
# print("Honeypot IoT-23 Shape " + str(honeypot.shape))
# print("Benign IoT-Benign-Trafic Shape " + str(benign_iottr.shape))
print("==========================================")
print("Total Malware " + str(malware.shape[0]))
print("Total Benign " + str(benign_iot23.shape[0]))
print("Total Dataset " + str(malware.shape[0]+benign_iot23.shape[0]))

# benign = pd.concat([benign_iot23, honeypot])
dataset = pd.concat([benign_iot23, malware])

le = preprocessing.LabelEncoder()
label = le.fit_transform(dataset["Label"])

malware["Label"] = 1
benign_iot23["Label"] = 0

benign_iot23["Label"]

dataset["Label"] = label

dataset["Label"]

x = dataset.drop(["Label"], axis=1)
y = dataset[["Label"]]
smp = RandomUnderSampler(random_state=0, sampling_strategy='majority')
x_smp, y_smp = smp.fit_resample(x, y)

x_smp.shape

y_smp.shape

malware.shape

dataset.shape

x = dataset.drop(["Label"], axis=1)
y = dataset[["Label"]]

# sampler = TomekLinks(sampling_strategy="all")
sampler = RandomUnderSampler(random_state=0)

x_resampled, y_resampled = sampler.fit_resample(x, y)

x_resampled.shape

y_resampled.shape

malsample = malware.sample(frac=0.0005, random_state=0)
bensample = benign.sample(frac=0.0005, random_state=0)

print(malsample.shape)
print(bensample.shape)
print("Total Dataset " + str(malsample.shape[0] + bensample.shape[0]))

dataset_sampled = pd.concat([malsample, bensample])

## Gabungkan Malware Dataset

malware.to_csv("/media/kmdr7/Seagate/TA/DATASETS/NewMalware.csv", index=False)

## Gabungkan Benign Dataset

benign_iot23.to_csv("/media/kmdr7/Seagate/TA/DATASETS/NewBenign.csv", index=False)

## Gabung Dataset

dataset.to_csv("/media/kmdr7/Seagate/TA/DATASETS/new/Dataset.csv", index=False)

dataset_sampled.to_csv("/media/kmdr7/Seagate/TA/DATASETS/DatasetSampledSmall.csv", index=False)
