==> CLEAN CICFLOWMETER/01. Data Preparation/01. Dataset Visualization TSNE 3D.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import plotly.express as px
import plotly.graph_objects as go # or plotly.express as px

from sklearn.manifold import TSNE
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.decomposition import PCA

import dash
import dash_core_components as dcc
import dash_html_components as html

dataset = pd.read_csv('/media/kmdr7/Seagate/TA/DATASETS/DatasetSampledSmall.csv')

dataset

x = dataset.drop(["Label"], axis=1)
y = dataset[["Label"]]

dataset.isnull().sum().sum()

tsne = TSNE(n_components=3, verbose=1, random_state=0)
z = tsne.fit_transform(x)

z

df = pd.DataFrame(z, columns=("comp-1", "comp-2", "comp-3"))
df['Label'] = y
df

fig = px.scatter_3d(df, x='comp-1', y='comp-2', z='comp-3',
              color='Label')

# fig.write_image("fig1.png")

app = dash.Dash()
app.layout = html.Div([
    dcc.Graph(figure=fig)
])

app.run_server(debug=True, use_reloader=False)  # Turn off reloader if inside Jupyter
