==> CLEAN CICFLOWMETER/01. Data Preparation/01. Dataset Visualization.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import plotly.express as px

from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.decomposition import PCA

pd.options.display.max_rows = 100

dataset = pd.read_csv('/media/kmdr7/Seagate/TA/DATASETS/DatasetSampled.csv')

dataset

dataset.info()

dataset.replace([np.inf, -np.inf], np.nan, inplace=True)

dataset.isnull().sum().sum()

dataset.interpolate(method ='linear', limit_direction ='forward', inplace=True)

features = dataset.columns.tolist()
features

fig = px.scatter_matrix(
    dataset,
    dimensions=features,
    color="Label"
)
fig.update_traces(diagonal_visible=False)
fig.show()

x = dataset.drop(columns=["Label"])
y = dataset.Label

std_scaler = StandardScaler()
pca = PCA(n_components=2)

plotds = pd.DataFrame(std_scaler.fit_transform(x), columns=x.columns)
plotds = pd.DataFrame(pca.fit_transform(x), columns=["a","b"])
plotds["Label"] = y

plotds

fig = plt.figure(figsize = (8,8))
ax = fig.add_subplot(1,1,1) 
ax.set_xlabel('X', fontsize = 15)
ax.set_ylabel('Y', fontsize = 15)
ax.set_title('Dataset Visualization', fontsize = 20)
targets = ['Benign', 'Malicious']
colors = ['r', 'g']
for target, color in zip(targets,colors):
    indicesToKeep = plotds['Label'] == target
    ax.scatter(plotds.loc[indicesToKeep, 'a']
               , plotds.loc[indicesToKeep, 'b']
               , c = color
               , s = 50)
ax.legend(targets)
ax.grid()

dataset.Label.unique()

# sns.catplot(x="Label", data=dataset, kind="count", aspect=2, legend=True)

dataset[dataset["Label"] == "Benign"].shape

dataset[dataset["Label"] == "Malicious"].shape
