==> CLEAN CICFLOWMETER/01. Data Preparation/02. Resampling Data.ipynb <==
import pandas as pd
from sklearn.model_selection import StratifiedKFold
from sklearn import preprocessing
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.ensemble import RandomForestClassifier, BaggingClassifier

le = preprocessing.LabelEncoder()

dataset = pd.read_csv("/media/kmdr7/Seagate/TA/DATASETS/new/Dataset.csv")

mal = dataset[dataset["Label"]==1]
ben = dataset[dataset["Label"]==0]

print(mal.shape)
print(ben.shape)

malsample = mal.head(1000000)
bensample = ben.head(1000000)

print(malsample.shape)
print(bensample.shape)

sample = pd.concat([malsample, bensample])

sample.shape

# label = le.fit_transform(sample["Label"])
# sample.drop(columns=["Label"], inplace=True)
# sample["Label"] = label

dataset.drop(sample.index, inplace=True)

dataset.shape

dataset.to_csv("/media/kmdr7/Seagate/TA/DATASETS/newUnseenDataset.csv", index=False)

sample.to_csv("/media/kmdr7/Seagate/TA/DATASETS/newDatasetSampledEncoded.csv", index=False)

lb = sample["Label"]
clean = sample.drop(columns=["Label"])

x_train, x_test, y_train, y_test = train_test_split(
    clean, lb, test_size=1 / 7.0, random_state=1
)

clf = RandomForestClassifier(random_state=0)
cross_val_score(estimator=clf, X=clean, y=lb, cv=10, n_jobs=8)
